#1503 implemented SmvCsvStringData  in iomod
